chore: 체험단 MY > 체험 페이지 API의 RequestDTO의 targetPage의 타입을 Long에서 Integer로 변경

+++ src/main/java/com/nuguna/freview/customer/dto/request/CustomerMyAcceptedApplyInfosRetrieveRequestDTO.java
@@ -13,5 +13,5 @@ public class CustomerMyAcceptedApplyInfosRetrieveRequestDTO {
 
   @Min(1)
   @NotNull
-  private Long targetPage;
+  private Integer targetPage;
 }
