+++ src/main/java/com/nuguna/freview/customer/dto/request/CustomerProposalToMeInfosRetrieveRequestDTO.java
@@ -13,5 +13,5 @@ public class CustomerProposalToMeInfosRetrieveRequestDTO {
 
   @Min(1)
   @NotNull
-  private Long targetPage;
+  private Integer targetPage;
 }
